[ANCHOR-217] Deposit end-2-end tests in Kotlin (#799)
1. Build the test framework that runs on profiles that can be defined to test different AP configurations.
2. Add docker-compose to GH actions.
3. Output test stdout in Gradle build so that we may see what went wrong.
4. Integrate Kotlin Wallet SDK
5. Implement deposit e2e test in Kotlin.
