+++ .github/workflows/sub_gradle_test_and_build.yml
@@ -10,30 +10,38 @@ jobs:
     name: Gradle Test and Build
     runs-on: ubuntu-22.04
     steps:
+      # Start the docker containers
       - uses: actions/checkout@v3
+      - name: Build the stack
+        env:
+          TEST_PROFILE_NAME: default
+        run: docker-compose -f integration-tests/src/test/resources/docker-compose-test.yaml up -d --build
+
+      # Set up JDK 11
       - name: Set up JDK 11
-        uses: actions/setup-java@v2
+        uses: actions/setup-java@v3
         with:
           java-version: '11'
           distribution: 'adopt'
 
-      - name: Spotless Check
-        run: ./gradlew spotlessCheck || echo "❌ Your code is not properly formatted. You can run './gradlew spotlessApply' to format it. 👀"
+      # Check the docker containers
+      - name: Check running containers
+        run: docker ps
 
-      - name: Test and Build
-        run: ./gradlew build
-
-      - name: Print Test Results
-        if: failure()
+      # This is to add to DNS entries to access the services started by docker-compose.
+      - name: Add 'db' and 'kafka' to /etc/hosts
         run: |
-          echo "\n\n*** API Schema test report ***\n"
-          cat /home/runner/work/java-stellar-anchor-sdk/java-stellar-anchor-sdk/api-schema/build/reports/tests/test/index.html
-          echo "\n\n*** SEP test report ***\n"
-          cat /home/runner/work/java-stellar-anchor-sdk/java-stellar-anchor-sdk/core/build/reports/tests/test/index.html
-          echo "\n\n*** Platform test report ***\n"
-          cat /home/runner/work/java-stellar-anchor-sdk/java-stellar-anchor-sdk/platform/build/reports/tests/test/index.html
-          echo "\n\n*** Integration tests report ***\n"
-          cat /home/runner/work/java-stellar-anchor-sdk/java-stellar-anchor-sdk/integration-tests/build/reports/tests/test/index.html
+          sudo echo "127.0.0.1 db" | sudo tee -a /etc/hosts
+          sudo echo "127.0.0.1 kafka" | sudo tee -a /etc/hosts
+          sudo nslookup db
+          sudo nslookup kafka
+
+      # Gradle test is now printing the test messages to GitHub Actions log. It is not necessary to print the reports.
+      - name: Test and Build
+        env:
+          TEST_PROFILE_NAME: default
+          run_docker: false
+        run: ./gradlew integration-tests:test --stacktrace
 
   analyze:
     name: CodeQL Analysis
