+++ integration-tests/build.gradle.kts
@@ -24,6 +24,9 @@ dependencies {
   implementation(libs.log4j2.core)
   implementation(libs.log4j2.slf4j)
   implementation(libs.docker.compose.rule)
+  implementation(libs.stellar.wallet.sdk)
+  implementation(libs.ktor.client.core)
+  implementation(libs.ktor.client.okhttp)
 
   // project dependencies
   implementation(project(":api-schema"))
