+++ integration-tests/src/test/kotlin/org/stellar/anchor/platform/ResourceHelper.kt
@@ -6,18 +6,27 @@ import java.io.File
 import java.net.URL
 
 fun readResourceAsMap(resourcePath: String): MutableMap<String, String> {
-  val resourceUrl: URL? = {}::class.java.classLoader.getResource(resourcePath)
-  val resourceFile = File(resourceUrl!!.toURI())
+  val resourceFilePath = getResourceFilePath(resourcePath)
+  return readResourceAsMap(File(resourceFilePath))
+}
 
+fun readResourceAsMap(resourceFile: File): MutableMap<String, String> {
   return DotenvParser(
-          DotenvReader(resourceFile.parentFile.absolutePath, resourceFile.name), false, false)
-      .parse()
-      .associate { it.key to it.value }
-      .toMutableMap()
+      DotenvReader(resourceFile.parentFile.absolutePath, resourceFile.name),
+      false,
+      false
+    )
+    .parse()
+    .associate { it.key to it.value }
+    .toMutableMap()
 }
 
 fun getResourceFilePath(resourcePath: String): String {
-  val resourceUrl: URL? = {}::class.java.classLoader.getResource(resourcePath)
+  val resourceUrl: URL? =
+    {}::class
+      .java
+      .classLoader
+      .getResource(if (resourcePath.startsWith("/")) resourcePath.substring(1) else resourcePath)
   val resourceFile = File(resourceUrl!!.toURI())
   return resourceFile.absolutePath
 }
