+++ integration-tests/src/test/kotlin/org/stellar/anchor/platform/TestAccounts.kt
@@ -1,0 +1,10 @@
+package org.stellar.anchor.platform
+
+var CLIENT_WALLET_SECRET = "SCQ2YRUGKBO4Q6RWSXV3SYVGHJB525XVKUP2HDGVW3LV42UBLGTZ24XX"
+var CLIENT_WALLET_ACCOUNT = "GCHU3RZAECOKGM2YAJLQIIYB2ZPLMFTTGN5D3XZNX4RDOEERVLXO7HU4"
+var CLIENT_WALLET_EXTRA_SIGNER_1_SECRET = "SDUDRKCL5AX7RDZ7S6JAPBNKLV6LRZXJSHN5OJDP32TIJB42ODPQODHY"
+var CLIENT_WALLET_EXTRA_SIGNER_1_ACCOUNT =
+  "GC6X2ANA2OS3O2ESHUV6X44NH6J46EP2EO2JB7563Y7DYOIXFKHMHJ5O"
+var CLIENT_WALLET_EXTRA_SIGNER_2_SECRET = "SC52GRNSIODLPNGTXUCZ5NHBII4PYUKUVQCCWWIK6OB6P4AW4M37DXZK"
+var CLIENT_WALLET_EXTRA_SIGNER_2_ACCOUNT =
+  "GATEYCIMJZ2F6Y437QSYH4XFQ6HLD5YP4MBJZFFPZVEQDJOY4QTCB7BB"
