+++ integration-tests/src/test/kotlin/org/stellar/anchor/platform/test/Sep24Tests.kt
@@ -1,6 +1,6 @@
 @file:Suppress("UNCHECKED_CAST")
 
-package org.stellar.anchor.platform
+package org.stellar.anchor.platform.test
 
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.assertThrows
@@ -10,140 +10,139 @@ import org.springframework.web.util.UriComponentsBuilder
 import org.stellar.anchor.api.exception.SepException
 import org.stellar.anchor.api.platform.PatchTransactionsRequest
 import org.stellar.anchor.api.sep.sep24.Sep24GetTransactionResponse
+import org.stellar.anchor.apiclient.PlatformApiClient
+import org.stellar.anchor.auth.AuthHelper
 import org.stellar.anchor.auth.JwtService
 import org.stellar.anchor.auth.Sep24InteractiveUrlJwt
 import org.stellar.anchor.auth.Sep24MoreInfoUrlJwt
-import org.stellar.anchor.platform.AnchorPlatformIntegrationTest.Companion.SEP10_JWT_SECRET
-import org.stellar.anchor.platform.AnchorPlatformIntegrationTest.Companion.SEP24_INTERACTIVE_URL_JWT_SECRET
-import org.stellar.anchor.platform.AnchorPlatformIntegrationTest.Companion.SEP24_MORE_INFO_URL_JWT_SECRET
-import org.stellar.anchor.platform.AnchorPlatformIntegrationTest.Companion.jwt
-import org.stellar.anchor.platform.AnchorPlatformIntegrationTest.Companion.toml
+import org.stellar.anchor.platform.*
+import org.stellar.anchor.util.Sep1Helper.*
 
-lateinit var sep24Client: Sep24Client
 lateinit var savedWithdrawTxn: Sep24GetTransactionResponse
 lateinit var savedDepositTxn: Sep24GetTransactionResponse
 
-class Sep24Tests {
-  companion object {
-    private val jwtService =
-      JwtService(SEP10_JWT_SECRET, SEP24_INTERACTIVE_URL_JWT_SECRET, SEP24_MORE_INFO_URL_JWT_SECRET)
-    fun setup() {
-      sep24Client = Sep24Client(toml.getString("TRANSFER_SERVER_SEP0024"), jwt)
-    }
-    fun `test Sep24 info endpoint`() {
-      printRequest("Calling GET /info")
-      val info = sep24Client.getInfo()
-      JSONAssert.assertEquals(expectedSep24Info, gson.toJson(info), LENIENT)
-    }
+class Sep24Tests(val config: TestConfig, val toml: TomlContent, jwt: String) {
+  private val jwtService: JwtService =
+    JwtService(
+      config.env["secret.sep10.jwt_secret"]!!,
+      config.env["secret.sep24.interactive_url.jwt_secret"]!!,
+      config.env["secret.sep24.more_info_url.jwt_secret"]!!
+    )
 
-    fun `test Sep24 withdraw`() {
-      printRequest("POST /transactions/withdraw/interactive")
-      val withdrawRequest = gson.fromJson(withdrawRequest, HashMap::class.java)
-      val response = sep24Client.withdraw(withdrawRequest as HashMap<String, String>)
-      printResponse("POST /transactions/withdraw/interactive response:", response)
-      savedWithdrawTxn = sep24Client.getTransaction(response.id, "USDC")
-      printResponse(savedWithdrawTxn)
-      JSONAssert.assertEquals(expectedSep24WithdrawResponse, json(savedWithdrawTxn), LENIENT)
-      // check the returning Sep24InteractiveUrlJwt
-      val params = UriComponentsBuilder.fromUriString(response.url).build().queryParams
-      val cipher = params["token"]!![0]
-      val jwt = jwtService.decode(cipher, Sep24InteractiveUrlJwt::class.java)
-      val data = jwt.claims["data"] as Map<String, String>
-      assertEquals(response.id, data["transaction_id"])
-    }
+  private val platformApiClient =
+    PlatformApiClient(AuthHelper.forNone(), config.env["platform.server.url"]!!)
+  private val sep24Client = Sep24Client(toml.getString("TRANSFER_SERVER_SEP0024"), jwt)
 
-    fun `test Sep24 deposit`() {
-      printRequest("POST /transactions/withdraw/interactive")
-      val depositRequest = gson.fromJson(depositRequest, HashMap::class.java)
-      val response = sep24Client.deposit(depositRequest as HashMap<String, String>)
-      printResponse("POST /transactions/deposit/interactive response:", response)
-      savedDepositTxn = sep24Client.getTransaction(response.id, "USDC")
-      printResponse(savedDepositTxn)
-      JSONAssert.assertEquals(expectedSep24DepositResponse, json(savedDepositTxn), LENIENT)
-      // check the returning Sep24InteractiveUrlJwt
-      val params = UriComponentsBuilder.fromUriString(response.url).build().queryParams
-      val cipher = params["token"]!![0]
-      val jwt = jwtService.decode(cipher, Sep24InteractiveUrlJwt::class.java)
-      val data = jwt.claims["data"] as Map<String, String>
-      assertEquals(response.id, data["transaction_id"])
-    }
+  private fun `test Sep24 info endpoint`() {
+    printRequest("Calling GET /info")
+    val info = sep24Client.getInfo()
+    JSONAssert.assertEquals(expectedSep24Info, gson.toJson(info), LENIENT)
+  }
 
-    fun `test Sep24 GET transaction and check the JWT`() {
-      val txn =
-        sep24Client
-          .getTransaction(
-            savedDepositTxn.transaction.id,
-            "stellar:USDC:GDQOE23CFSUMSVQK4Y5JHPPYK73VYCNHZHA7ENKCV37P6SUEO6XQBKPP"
-          )
-          .transaction
+  private fun `test Sep24 withdraw`() {
+    printRequest("POST /transactions/withdraw/interactive")
+    val withdrawRequest = gson.fromJson(withdrawRequest, HashMap::class.java)
+    val response = sep24Client.withdraw(withdrawRequest as HashMap<String, String>)
+    printResponse("POST /transactions/withdraw/interactive response:", response)
+    savedWithdrawTxn = sep24Client.getTransaction(response.id, "USDC")
+    printResponse(savedWithdrawTxn)
+    JSONAssert.assertEquals(expectedSep24WithdrawResponse, json(savedWithdrawTxn), LENIENT)
+    // check the returning Sep24InteractiveUrlJwt
+    val params = UriComponentsBuilder.fromUriString(response.url).build().queryParams
+    val cipher = params["token"]!![0]
+    val jwt = jwtService.decode(cipher, Sep24InteractiveUrlJwt::class.java)
+    val data = jwt.claims["data"] as Map<String, String>
+    assertEquals(response.id, data["transaction_id"])
+  }
 
-      val params = UriComponentsBuilder.fromUriString(txn.moreInfoUrl).build().queryParams
-      val cipher = params["token"]!![0]
-      val jwt = jwtService.decode(cipher, Sep24MoreInfoUrlJwt::class.java)
-      val data = jwt.claims["data"] as Map<String, String>
-      assertEquals(txn.id, data["transaction_id"])
-    }
+  private fun `test Sep24 deposit`() {
+    printRequest("POST /transactions/withdraw/interactive")
+    val depositRequest = gson.fromJson(depositRequest, HashMap::class.java)
+    val response = sep24Client.deposit(depositRequest as HashMap<String, String>)
+    printResponse("POST /transactions/deposit/interactive response:", response)
+    savedDepositTxn = sep24Client.getTransaction(response.id, "USDC")
+    printResponse(savedDepositTxn)
+    JSONAssert.assertEquals(expectedSep24DepositResponse, json(savedDepositTxn), LENIENT)
+    // check the returning Sep24InteractiveUrlJwt
+    val params = UriComponentsBuilder.fromUriString(response.url).build().queryParams
+    val cipher = params["token"]!![0]
+    val jwt = jwtService.decode(cipher, Sep24InteractiveUrlJwt::class.java)
+    val data = jwt.claims["data"] as Map<String, String>
+    assertEquals(response.id, data["transaction_id"])
+  }
 
-    fun `test PlatformAPI GET transaction for deposit and withdrawal`() {
-      val actualWithdrawTxn = platformApiClient.getTransaction(savedWithdrawTxn.transaction.id)
-      assertEquals(actualWithdrawTxn.id, savedWithdrawTxn.transaction.id)
-      println(expectedWithdrawTransactionResponse)
-      println(json(actualWithdrawTxn))
-      JSONAssert.assertEquals(expectedWithdrawTransactionResponse, json(actualWithdrawTxn), LENIENT)
+  private fun `test Sep24 GET transaction and check the JWT`() {
+    val txn =
+      sep24Client
+        .getTransaction(
+          savedDepositTxn.transaction.id,
+          "stellar:USDC:GDQOE23CFSUMSVQK4Y5JHPPYK73VYCNHZHA7ENKCV37P6SUEO6XQBKPP"
+        )
+        .transaction
 
-      val actualDepositTxn = platformApiClient.getTransaction(savedDepositTxn.transaction.id)
-      printResponse(actualDepositTxn)
-      assertEquals(actualDepositTxn.id, savedDepositTxn.transaction.id)
-      JSONAssert.assertEquals(expectedDepositTransactionResponse, json(actualDepositTxn), LENIENT)
-    }
+    val params = UriComponentsBuilder.fromUriString(txn.moreInfoUrl).build().queryParams
+    val cipher = params["token"]!![0]
+    val jwt = jwtService.decode(cipher, Sep24MoreInfoUrlJwt::class.java)
+    val data = jwt.claims["data"] as Map<String, String>
+    assertEquals(txn.id, data["transaction_id"])
+  }
 
-    fun `test patch, get and compare`() {
-      val patch =
-        gson.fromJson(patchWithdrawTransactionRequest, PatchTransactionsRequest::class.java)
-      // create patch request and patch
-      patch.records[0].transaction.id = savedWithdrawTxn.transaction.id
-      patch.records[1].transaction.id = savedDepositTxn.transaction.id
-      platformApiClient.patchTransaction(patch)
+  private fun `test PlatformAPI GET transaction for deposit and withdrawal`() {
+    val actualWithdrawTxn = platformApiClient.getTransaction(savedWithdrawTxn.transaction.id)
+    assertEquals(actualWithdrawTxn.id, savedWithdrawTxn.transaction.id)
+    println(expectedWithdrawTransactionResponse)
+    println(json(actualWithdrawTxn))
+    JSONAssert.assertEquals(expectedWithdrawTransactionResponse, json(actualWithdrawTxn), LENIENT)
 
-      // check if the patched transactions are as expected
-      var afterPatchWithdraw = platformApiClient.getTransaction(savedWithdrawTxn.transaction.id)
-      assertEquals(afterPatchWithdraw.id, savedWithdrawTxn.transaction.id)
-      JSONAssert.assertEquals(expectedAfterPatchWithdraw, json(afterPatchWithdraw), LENIENT)
+    val actualDepositTxn = platformApiClient.getTransaction(savedDepositTxn.transaction.id)
+    printResponse(actualDepositTxn)
+    assertEquals(actualDepositTxn.id, savedDepositTxn.transaction.id)
+    JSONAssert.assertEquals(expectedDepositTransactionResponse, json(actualDepositTxn), LENIENT)
+  }
 
-      var afterPatchDeposit = platformApiClient.getTransaction(savedDepositTxn.transaction.id)
-      assertEquals(afterPatchDeposit.id, savedDepositTxn.transaction.id)
-      JSONAssert.assertEquals(expectedAfterPatchDeposit, json(afterPatchDeposit), LENIENT)
+  private fun `test patch, get and compare`() {
+    val patch = gson.fromJson(patchWithdrawTransactionRequest, PatchTransactionsRequest::class.java)
+    // create patch request and patch
+    patch.records[0].transaction.id = savedWithdrawTxn.transaction.id
+    patch.records[1].transaction.id = savedDepositTxn.transaction.id
+    platformApiClient.patchTransaction(patch)
 
-      // Test patch idempotency
-      afterPatchWithdraw = platformApiClient.getTransaction(savedWithdrawTxn.transaction.id)
-      assertEquals(afterPatchWithdraw.id, savedWithdrawTxn.transaction.id)
-      JSONAssert.assertEquals(expectedAfterPatchWithdraw, json(afterPatchWithdraw), LENIENT)
+    // check if the patched transactions are as expected
+    var afterPatchWithdraw = platformApiClient.getTransaction(savedWithdrawTxn.transaction.id)
+    assertEquals(afterPatchWithdraw.id, savedWithdrawTxn.transaction.id)
+    JSONAssert.assertEquals(expectedAfterPatchWithdraw, json(afterPatchWithdraw), LENIENT)
 
-      afterPatchDeposit = platformApiClient.getTransaction(savedDepositTxn.transaction.id)
-      assertEquals(afterPatchDeposit.id, savedDepositTxn.transaction.id)
-      JSONAssert.assertEquals(expectedAfterPatchDeposit, json(afterPatchDeposit), LENIENT)
-    }
+    var afterPatchDeposit = platformApiClient.getTransaction(savedDepositTxn.transaction.id)
+    assertEquals(afterPatchDeposit.id, savedDepositTxn.transaction.id)
+    JSONAssert.assertEquals(expectedAfterPatchDeposit, json(afterPatchDeposit), LENIENT)
 
-    fun `test GET transactions with bad ids`() {
-      val badTxnIds = listOf("null", "bad id", "123", null)
-      for (txnId in badTxnIds) {
-        assertThrows<SepException> { platformApiClient.getTransaction(txnId) }
-      }
+    // Test patch idempotency
+    afterPatchWithdraw = platformApiClient.getTransaction(savedWithdrawTxn.transaction.id)
+    assertEquals(afterPatchWithdraw.id, savedWithdrawTxn.transaction.id)
+    JSONAssert.assertEquals(expectedAfterPatchWithdraw, json(afterPatchWithdraw), LENIENT)
+
+    afterPatchDeposit = platformApiClient.getTransaction(savedDepositTxn.transaction.id)
+    assertEquals(afterPatchDeposit.id, savedDepositTxn.transaction.id)
+    JSONAssert.assertEquals(expectedAfterPatchDeposit, json(afterPatchDeposit), LENIENT)
+  }
+
+  private fun `test GET transactions with bad ids`() {
+    val badTxnIds = listOf("null", "bad id", "123", null)
+    for (txnId in badTxnIds) {
+      assertThrows<SepException> { platformApiClient.getTransaction(txnId) }
     }
   }
-}
 
-fun sep24TestAll() {
-  Sep24Tests.setup()
-
-  println("Performing SEP24 tests...")
-  Sep24Tests.`test Sep24 info endpoint`()
-  Sep24Tests.`test Sep24 withdraw`()
-  Sep24Tests.`test Sep24 deposit`()
-  Sep24Tests.`test Sep24 GET transaction and check the JWT`()
-  Sep24Tests.`test PlatformAPI GET transaction for deposit and withdrawal`()
-  Sep24Tests.`test patch, get and compare`()
-  Sep24Tests.`test GET transactions with bad ids`()
+  fun testAll() {
+    println("Performing SEP24 tests...")
+    `test Sep24 info endpoint`()
+    `test Sep24 withdraw`()
+    `test Sep24 deposit`()
+    `test Sep24 GET transaction and check the JWT`()
+    `test PlatformAPI GET transaction for deposit and withdrawal`()
+    `test patch, get and compare`()
+    `test GET transactions with bad ids`()
+  }
 }
 
 private const val withdrawRequest =
@@ -158,7 +157,7 @@ private const val depositRequest =
   """{
     "asset_code": "USDC",
     "asset_issuer": "GDQOE23CFSUMSVQK4Y5JHPPYK73VYCNHZHA7ENKCV37P6SUEO6XQBKPP",
-    "account": "GDJLBYYKMCXNVVNABOE66NYXQGIA5AC5D223Z2KF6ZEYK4UBCA7FKLTG",
+    "account": "GCHU3RZAECOKGM2YAJLQIIYB2ZPLMFTTGN5D3XZNX4RDOEERVLXO7HU4",
     "lang": "en"
 }"""
 
@@ -385,7 +384,7 @@ private const val expectedSep24DepositResponse =
       "kind": "deposit",
       "status": "incomplete",
       "refunded": false,
-      "to": "GDJLBYYKMCXNVVNABOE66NYXQGIA5AC5D223Z2KF6ZEYK4UBCA7FKLTG"
+      "to": "GCHU3RZAECOKGM2YAJLQIIYB2ZPLMFTTGN5D3XZNX4RDOEERVLXO7HU4"
     }
   }
 """
