+++ integration-tests/src/test/resources/config/assets.yaml
