+++ integration-tests/src/test/resources/config/java-reference-server-config.yaml
