+++ integration-tests/src/test/resources/docker-compose-test.yaml
@@ -1,19 +1,22 @@
 version: '2.4'
 services:
   kafka:
+    hostname: kafka
     extends:
-      file: ../common/docker-compose.yaml
+      file: common/docker-compose.yaml
       service: kafka
     depends_on:
       - zookeeper
 
   zookeeper:
+    hostname: zookeeper
     extends:
-      file: ../common/docker-compose.yaml
+      file: common/docker-compose.yaml
       service: zookeeper
 
   db:
+    hostname: db
     extends:
-      file: ../common/docker-compose.yaml
+      file: common/docker-compose.yaml
       service: db
