+++ integration-tests/src/test/resources/profiles/default/config.env
@@ -12,10 +12,10 @@ logging.stellar_level=DEBUG
 # events
 events.enabled=true
 events.publisher.type=kafka
-events.publisher.kafka.bootstrap_server=localhost:29092
+events.publisher.kafka.bootstrap_server=kafka:29092
 # data
 data.type=postgres
-data.server=localhost:5432
+data.server=db:5432
 data.database=postgres
 data.flyway_enabled=true
 # assets
@@ -30,7 +30,6 @@ sep10.home_domain=localhost:8080
 sep12.enabled=true
 sep31.enabled=true
 sep38.enabled=true
-
 sep24.enabled=true
-sep24.interactive_url.base_url=https://stellar.moneygram.com
+sep24.interactive_url.base_url=http://localhost:8091/sep24/interactive
 sep24.more_info_url.base_url=http://localhost:8080/sep24/transaction/more_info
