+++ integration-tests/src/test/resources/profiles/sep24/config.env
@@ -1,0 +1,35 @@
+# secrets
+secret.data.username=postgres
+secret.data.password=password
+secret.platform_api.auth_secret=myAnchorToPlatformSecret
+secret.callback_api.auth_secret=myPlatformToAnchorSecret
+secret.sep10.jwt_secret=secret_sep10_secret
+secret.sep10.signing_seed=SAX3AH622R2XT6DXWWSRIDCMMUCCMATBZ5U6XKJWDO7M2EJUBFC3AW5X
+secret.sep24.interactive_url.jwt_secret=secret_sep24_interactive_url_jwt_secret
+secret.sep24.more_info_url.jwt_secret=secret_sep24_more_info_url_jwt_secret
+# logging
+logging.stellar_level=DEBUG
+# events
+events.enabled=true
+events.publisher.type=kafka
+events.publisher.kafka.bootstrap_server=localhost:29092
+# data
+data.type=postgres
+data.server=localhost:5432
+data.database=postgres
+data.flyway_enabled=true
+# assets
+assets.type=file
+assets.value=/config/assets.yaml
+# seps
+sep1.enabled=true
+sep1.toml.type=file
+sep1.toml.value=/config/stellar.toml
+sep10.enabled=true
+sep10.home_domain=localhost:8080
+sep12.enabled=true
+sep31.enabled=true
+sep38.enabled=true
+sep24.enabled=true
+sep24.interactive_url.base_url=http://localhost:8091/sep24/interactive
+sep24.more_info_url.base_url=http://localhost:8080/sep24/transaction/more_info
