+++ integration-tests/src/test/resources/profiles/sep24/test.env
@@ -1,0 +1,11 @@
+# Test configurations
+anchor.domain=http://localhost:8080
+reference.server.url=http://localhost:8089
+platform.server.url=http://localhost:8080
+observer.server.url=http://localhost:8083
+run_docker=true
+run_servers=true
+
+# Enable the test flows in the kotlin reference server
+sep24.enableTest=true
+sep24.interactiveJwtKey=secret_sep24_interactive_url_jwt_secret
