+++ kotlin-reference-server/src/main/kotlin/org/stellar/reference/RefenreceServerStart.kt
@@ -2,6 +2,7 @@ package org.stellar.reference
 
 // Used in ServiceRunner
 @JvmOverloads // Annotation required to call from Java with optional argument
-fun start(waitServer: Boolean = false) = startServer(waitServer)
-fun stop() = stopServer()
+fun start(envMap: Map<String, String>, waitServer: Boolean = false) =
+  startServer(envMap, waitServer)
 
+fun stop() = stopServer()
