--- kotlin-reference-server/src/main/resources/docker-config.yaml
@@ -1,2 +1,0 @@
-sep24:
-  port: 8092
