+++ service-runner/src/main/java/org/stellar/anchor/platform/ServiceRunner.java
@@ -46,7 +46,7 @@ public class ServiceRunner {
       }
 
       if (cmd.hasOption("kotlin-reference-server") || cmd.hasOption("all")) {
-        startKotlinReferenceServer(true);
+        startKotlinReferenceServer(null, true);
         anyServerStarted = true;
       }
 
@@ -66,7 +66,6 @@ public class ServiceRunner {
     return new StellarObservingServer().start(env);
   }
 
-
   public static ConfigurableApplicationContext startEventProcessor(Map<String, String> env) {
     return new EventProcessingServer().start(env);
   }
@@ -83,11 +82,9 @@ public class ServiceRunner {
     return AnchorReferenceServer.start(port, "/");
   }
 
-
-  public static void startKotlinReferenceServer(boolean wait) {
-    RefenreceServerStartKt.start(wait);
+  public static void startKotlinReferenceServer(Map<String, String> envMap, boolean wait) {
+    RefenreceServerStartKt.start(envMap, wait);
   }
-
 
   static void printUsage(Options options) {
     HelpFormatter helper = new HelpFormatter();
